Move ReturnSequence to Functions namespace
